enh: Move point set I/O functions from PointSet to IO module [PointSet -> IO]
Closes #172.
