+++ Applications/src/calculate-surface-attributes.cc
@@ -20,6 +20,7 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
+#include "mirtk/PointSetIO.h"
 #include "mirtk/PointSetUtils.h"
 #include "mirtk/PolyDataCurvature.h"
 #include "mirtk/PolyDataSmoothing.h"
