+++ Applications/src/copy-pointset-attributes.cc
@@ -20,6 +20,7 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
+#include "mirtk/PointSetIO.h"
 #include "mirtk/PointSetUtils.h"
 
 #include "vtkSmartPointer.h"
