+++ Applications/src/decimate-surface.cc
@@ -20,7 +20,8 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
-#include "mirtk/PointSetUtils.h"
+#include "mirtk/PointSetIO.h"
+#include "mirtk/Vtk.h"
 
 #include "vtkPolyData.h"
 #include "vtkPolyDataReader.h"
