+++ Applications/src/evaluate-surface-overlap.cc
@@ -27,6 +27,7 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
+#include "mirtk/PointSetIO.h"
 #include "mirtk/PointSetUtils.h"
 
 #include "vtkCell.h"
