+++ Applications/src/extract-connected-points.cc
@@ -20,7 +20,8 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
-#include "mirtk/PointSetUtils.h"
+#include "mirtk/Vtk.h"
+#include "mirtk/PointSetIO.h"
 
 #include "vtkSmartPointer.h"
 #include "vtkPolyData.h"
