+++ Applications/src/offset-surface.cc
@@ -20,8 +20,11 @@
 #include "mirtk/Common.h"
 #include "mirtk/Options.h"
 
+#include "mirtk/PointSetIO.h"
 #include "mirtk/PointSetUtils.h"
 #include "mirtk/PolyDataSmoothing.h"
+
+#include "mirtk/Vtk.h"
 #include "mirtk/VtkMath.h"
 
 #include "vtkSmartPointer.h"
