+++ Applications/src/project-onto-surface.cc
@@ -23,10 +23,12 @@
 #include "mirtk/Vector.h"
 #include "mirtk/Matrix.h"
 #include "mirtk/GenericImage.h"
-#include "mirtk/PointSetUtils.h"
+#include "mirtk/PointSetIO.h"
 #include "mirtk/NearestNeighborInterpolateImageFunction.h"
 #include "mirtk/LinearInterpolateImageFunction.h"
 #include "mirtk/EuclideanDistanceTransform.h"
+
+#include "mirtk/Vtk.h"
 
 #include "vtkGenericCell.h"
 #include "vtkPolyData.h"
@@ -45,16 +47,6 @@
 #include "vtkCellLocator.h"
 #include "vtkOBBTree.h"
 #include "vtkModifiedBSPTree.h"
-
-#include "vtkPolyDataReader.h"
-#include "vtkXMLPolyDataReader.h"
-#include "vtkOBJReader.h"
-#include "vtkSTLReader.h"
-#include "vtkSTLWriter.h"
-#include "vtkPLYReader.h"
-#include "vtkPLYWriter.h"
-#include "vtkXMLPolyDataWriter.h"
-#include "vtkPolyDataWriter.h"
 
 using namespace mirtk;
