+++ Modules/IO/config/Config.cmake.in
@@ -58,6 +58,9 @@ basis_set_config(WITH_NIfTI @NiftiCLib_FOUND@)
 ## @brief Whether this library was built with PNG support
 basis_set_config(WITH_PNG @PNG_FOUND@)
 
+## @brief Whether this library was built with VTK support
+basis_set_config(WITH_VTK @VTK_FOUND@)
+
 ## @brief Whether this library was built with ZLIB support
 basis_set_config(WITH_ZLIB @ZLIB_FOUND@)
