+++ Modules/PointSet/src/PointSetUtils.cc
@@ -24,7 +24,6 @@
 #include "mirtk/Math.h"
 #include "mirtk/Memory.h"
 #include "mirtk/Stream.h"
-#include "mirtk/Path.h"
 #include "mirtk/Parallel.h"
 #include "mirtk/EdgeTable.h"
 #include "mirtk/DataStatistics.h"
@@ -36,7 +35,6 @@
 #include "mirtk/Vector3.h"
 #include "mirtk/Algorithm.h"
 
-#include "vtkSmartPointer.h"
 #include "vtkCellArray.h"
 #include "vtkPolyData.h"
 #include "vtkUnstructuredGrid.h"
@@ -54,20 +52,6 @@
 #include "vtkImageData.h"
 #include "vtkImageStencilData.h"
 
-#include "vtkOBJReader.h"
-#include "vtkSTLReader.h"
-#include "vtkSTLWriter.h"
-#include "vtkPLYReader.h"
-#include "vtkPLYWriter.h"
-#include "vtkXMLPolyDataReader.h"
-#include "vtkXMLPolyDataWriter.h"
-#include "vtkGenericDataObjectReader.h"
-#include "vtkXMLGenericDataObjectReader.h"
-#include "vtkXMLDataSetWriter.h"
-#include "vtkDataSetWriter.h"
-#include "vtkPolyDataReader.h"
-#include "vtkPolyDataWriter.h"
-
 #include "vtkImageStencilIterator.h"
 #include "vtkPolyDataToImageStencil.h"
 #include "vtkCleanPolyData.h"
@@ -78,640 +62,11 @@
 #include "vtkDelaunay3D.h"
 #include "vtkMassProperties.h"
 
-#include "dfsurface.h" // BrainSuite surface (.dfs) I/O functions
-
 #include <algorithm>
 
 
 namespace mirtk {
 
-
-// =============================================================================
-// I/O
-// =============================================================================
-
-// -----------------------------------------------------------------------------
-const char *DefaultExtension(vtkDataSet *dataset)
-{
-  if      (vtkPolyData        ::SafeDownCast(dataset)) return ".vtp";
-  else if (vtkUnstructuredGrid::SafeDownCast(dataset)) return ".vtu";
-  else if (vtkStructuredGrid  ::SafeDownCast(dataset)) return ".vts";
-  return ".vtk";
-}
-
-// -----------------------------------------------------------------------------
-vtkSmartPointer<vtkPointSet> ReadPointSet(const char *fname, int *ftype, bool exit_on_failure)
-{
-  vtkSmartPointer<vtkPointSet> pointset;
-  const string ext = Extension(fname);
-  if (ext == ".vtp" || ext == ".stl" || ext == ".ply" || ext == ".obj" || ext == ".dfs" || ext == ".off") {
-    pointset = ReadPolyData(fname);
-  } else if (ext.length() == 4  && ext.substr(0, 3) == ".vt" && ext != ".vtk") {
-    vtkSmartPointer<vtkXMLGenericDataObjectReader> reader;
-    reader = vtkSmartPointer<vtkXMLGenericDataObjectReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    pointset = vtkPointSet::SafeDownCast(reader->GetOutput());
-  } else {
-    vtkSmartPointer<vtkGenericDataObjectReader> reader;
-    reader = vtkSmartPointer<vtkGenericDataObjectReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    if (ftype) *ftype = reader->GetFileType();
-    pointset = vtkPointSet::SafeDownCast(reader->GetOutput());
-  }
-  if (exit_on_failure && (!pointset || pointset->GetNumberOfPoints() == 0)) {
-    cerr << "File " << fname << " either contains no points or could not be read" << endl;
-    exit(1);
-  }
-  return pointset;
-}
-
-// -----------------------------------------------------------------------------
-/// Read BrainSuite surface from .dfs file
-vtkSmartPointer<vtkPolyData> ReadDFS(const char *fname)
-{
-  vtkSmartPointer<vtkPolyData> polydata = vtkSmartPointer<vtkPolyData>::New();
-  // Read .dfs file
-  SILT::DFSurface surface;
-  if (!surface.readDFS(fname)) return polydata;
-  const vtkIdType npoints = static_cast<vtkIdType>(surface.vertices .size());
-  const vtkIdType ncells  = static_cast<vtkIdType>(surface.triangles.size());
-  // Copy vertex coordinates
-  vtkSmartPointer<vtkPoints> points = vtkSmartPointer<vtkPoints>::New();
-  points->SetNumberOfPoints(npoints);
-  double p[3];
-  for (vtkIdType i = 0; i < npoints; ++i) {
-    p[0] = static_cast<double>(surface.vertices[i].x);
-    p[1] = static_cast<double>(surface.vertices[i].y);
-    p[2] = static_cast<double>(surface.vertices[i].z);
-    points->SetPoint(i, p);
-  }
-  polydata->SetPoints(points);
-  // Copy triangle face list
-  vtkIdType pts[3];
-  vtkSmartPointer<vtkCellArray> cells = vtkSmartPointer<vtkCellArray>::New();
-  cells->Allocate(cells->EstimateSize(ncells, 3));
-  for (vtkIdType i = 0; i < ncells; ++i) {
-    pts[0] = static_cast<vtkIdType>(surface.triangles[i].a);
-    pts[1] = static_cast<vtkIdType>(surface.triangles[i].b);
-    pts[2] = static_cast<vtkIdType>(surface.triangles[i].c);
-    cells->InsertNextCell(3, pts);
-  }
-  polydata->SetPolys(cells);
-  // Copy vertex normals
-  if (!surface.vertexNormals.empty()) {
-    double n[3];
-    vtkSmartPointer<vtkFloatArray> normals = vtkSmartPointer<vtkFloatArray>::New();
-    normals->SetName("Normals");
-    normals->SetNumberOfComponents(3);
-    normals->SetNumberOfTuples(npoints);
-    for (vtkIdType i = 0; i < npoints; ++i) {
-      n[0] = static_cast<double>(surface.vertexNormals[i].x);
-      n[1] = static_cast<double>(surface.vertexNormals[i].y);
-      n[2] = static_cast<double>(surface.vertexNormals[i].z);
-      normals->SetTuple(i, n);
-    }
-    polydata->GetPointData()->SetNormals(normals);
-  }
-  // Copy vertex colors
-  if (!surface.vertexColors.empty()) {
-    double rgb[3];
-    vtkSmartPointer<vtkFloatArray> colors = vtkSmartPointer<vtkFloatArray>::New();
-    colors->SetName("Colors");
-    colors->SetNumberOfComponents(3);
-    colors->SetNumberOfTuples(npoints);
-    for (vtkIdType i = 0; i < npoints; ++i) {
-      rgb[0] = static_cast<double>(surface.vertexColors[i].x);
-      rgb[1] = static_cast<double>(surface.vertexColors[i].y);
-      rgb[2] = static_cast<double>(surface.vertexColors[i].z);
-      colors->SetTuple(i, rgb);
-    }
-    polydata->GetPointData()->AddArray(colors);
-  }
-  // Copy vertex UV coordinates
-  if (!surface.vertexUV.empty()) {
-    double uv[3] = {.0};
-    vtkSmartPointer<vtkFloatArray> coords = vtkSmartPointer<vtkFloatArray>::New();
-    coords->SetName("UV");
-    coords->SetNumberOfComponents(3);
-    coords->SetNumberOfTuples(npoints);
-    for (vtkIdType i = 0; i < npoints; ++i) {
-      uv[0] = static_cast<double>(surface.vertexUV[i].u);
-      uv[1] = static_cast<double>(surface.vertexUV[i].v);
-      coords->SetTuple(i, uv);
-    }
-    polydata->GetPointData()->SetTCoords(coords);
-  }
-  // Copy vertex labels
-  if (!surface.vertexLabels.empty()) {
-    vtkSmartPointer<vtkUnsignedShortArray> labels = vtkSmartPointer<vtkUnsignedShortArray>::New();
-    labels->SetName("Labels");
-    labels->SetNumberOfComponents(1);
-    labels->SetNumberOfTuples(npoints);
-    for (vtkIdType i = 0; i < npoints; ++i) {
-      labels->SetValue(i, surface.vertexLabels[i]);
-    }
-    polydata->GetPointData()->AddArray(labels);
-  }
-  // Copy vertex attributes
-  if (!surface.vertexAttributes.empty()) {
-    vtkSmartPointer<vtkFloatArray> scalars = vtkSmartPointer<vtkFloatArray>::New();
-    scalars->SetName("Attributes");
-    scalars->SetNumberOfComponents(1);
-    scalars->SetNumberOfTuples(npoints);
-    for (vtkIdType i = 0; i < npoints; ++i) {
-      scalars->SetValue(i, surface.vertexAttributes[i]);
-    }
-    polydata->GetPointData()->SetScalars(scalars);
-  }
-  return polydata;
-}
-
-// -----------------------------------------------------------------------------
-vtkSmartPointer<vtkPolyData> ReadOFF(const char *fname)
-{
-  vtkSmartPointer<vtkPolyData> polydata = vtkSmartPointer<vtkPolyData>::New();
-
-  ifstream ifs(fname);
-  if (!ifs) return polydata;
-
-  string keyword;
-  ifs >> keyword;
-  if (keyword != "OFF" && keyword != "off") return polydata;
-
-  int numVertices = -1, numFaces = -1, numEdges = -1;
-  ifs >> numVertices >> numFaces >> numEdges;
-  if (ifs.fail()) return polydata;
-
-  if (numVertices < 0) numVertices = 0;
-  if (numFaces    < 0) numFaces    = 0;
-
-  vtkSmartPointer<vtkPoints> points;
-  points = vtkSmartPointer<vtkPoints>::New();
-  points->SetNumberOfPoints(numVertices);
-
-  double p[3];
-  for (int i = 0; i < numVertices; ++i) {
-    ifs >> p[0] >> p[1] >> p[2];
-    if (ifs.fail()) break;
-    points->SetPoint(i, p);
-  }
-  if (ifs.fail()) return polydata;
-
-  vtkSmartPointer<vtkCellArray> verts, lines, polys;
-  vtkSmartPointer<vtkIdList> cell = vtkSmartPointer<vtkIdList>::New();
-  verts = vtkSmartPointer<vtkCellArray>::New();
-  lines = vtkSmartPointer<vtkCellArray>::New();
-  polys = vtkSmartPointer<vtkCellArray>::New();
-  for (int i = 0, ptId, n; i < numFaces; ++i) {
-    ifs >> n;
-    if (ifs.fail()) break;
-    if (n > 0) {
-      cell->Reset();
-      for (int j = 0; j < n; ++j) {
-        ifs >> ptId;
-        cell->InsertNextId(ptId);
-      }
-      if      (n == 1) verts->InsertNextCell(cell);
-      else if (n == 2) lines->InsertNextCell(cell);
-      else             polys->InsertNextCell(cell);
-      if (!ifs.good()) break;
-    }
-  }
-  if (ifs.fail()) return polydata;
-
-  verts->Squeeze();
-  lines->Squeeze();
-  polys->Squeeze();
-
-  polydata->SetPoints(points);
-  if (verts->GetNumberOfCells() > 0) polydata->SetVerts(verts);
-  if (lines->GetNumberOfCells() > 0) polydata->SetLines(lines);
-  if (polys->GetNumberOfCells() > 0) polydata->SetPolys(polys);
-
-  return polydata;
-}
-
-// -----------------------------------------------------------------------------
-vtkSmartPointer<vtkPolyData> ReadPolyData(const char *fname, int *ftype, bool exit_on_failure)
-{
-  vtkSmartPointer<vtkPolyData> polydata;
-  const string ext = Extension(fname);
-  if (ext == ".vtp") {
-    vtkSmartPointer<vtkXMLPolyDataReader> reader;
-    reader = vtkSmartPointer<vtkXMLPolyDataReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    polydata = reader->GetOutput();
-  } else if (ext == ".stl") {
-    vtkSmartPointer<vtkSTLReader> reader;
-    reader = vtkSmartPointer<vtkSTLReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    polydata = reader->GetOutput();
-  } else if (ext == ".ply") {
-    vtkSmartPointer<vtkPLYReader> reader;
-    reader = vtkSmartPointer<vtkPLYReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    polydata = reader->GetOutput();
-  } else if (ext == ".obj") {
-    vtkSmartPointer<vtkOBJReader> reader;
-    reader = vtkSmartPointer<vtkOBJReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    polydata = reader->GetOutput();
-  } else if (ext == ".dfs") {
-    polydata = ReadDFS(fname);
-  } else if (ext == ".off") {
-    polydata = ReadOFF(fname);
-  } else {
-    vtkSmartPointer<vtkPolyDataReader> reader;
-    reader = vtkSmartPointer<vtkPolyDataReader>::New();
-    reader->SetFileName(fname);
-    reader->Update();
-    if (ftype) *ftype = reader->GetFileType();
-    polydata = reader->GetOutput();
-  }
-  if (exit_on_failure && polydata->GetNumberOfPoints() == 0) {
-    cerr << "File " << fname << " either contains no points or could not be read" << endl;
-    exit(1);
-  }
-  return polydata;
-}
-
-// -----------------------------------------------------------------------------
-/// Write surface in BrainSuite .dfs format
-int WriteDFS(const char *fname, vtkPolyData *polydata)
-{
-  // Ensure surface is triangular mesh
-  vtkSmartPointer<vtkPolyData> mesh = Triangulate(polydata);
-  SILT::DFSurface surface;
-  // Copy vertex coordinates
-  double p[3];
-  surface.vertices.resize(mesh->GetNumberOfPoints());
-  for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-    mesh->GetPoint(i, p);
-    surface.vertices[i].x = static_cast<float>(p[0]);
-    surface.vertices[i].y = static_cast<float>(p[1]);
-    surface.vertices[i].z = static_cast<float>(p[2]);
-  }
-  // Copy triangular faces
-  vtkIdType npts, *pts;
-  surface.triangles.resize(mesh->GetNumberOfCells());
-  for (vtkIdType i = 0; i < mesh->GetNumberOfCells(); ++i) {
-    mesh->GetCellPoints(i, npts, pts);
-    if (npts != 3) return false;
-    surface.triangles[i].a = static_cast<int>(pts[0]);
-    surface.triangles[i].b = static_cast<int>(pts[1]);
-    surface.triangles[i].c = static_cast<int>(pts[2]);
-  }
-  // Copy vertex normals
-  vtkDataArray *normals = mesh->GetPointData()->GetNormals();
-  if (!normals) normals = mesh->GetPointData()->GetArray("Normals");
-  if (normals) {
-    double n[3];
-    surface.vertexNormals.resize(mesh->GetNumberOfPoints());
-    for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-      normals->GetTuple(i, n);
-      surface.vertexNormals[i].x = static_cast<float>(n[0]);
-      surface.vertexNormals[i].y = static_cast<float>(n[1]);
-      surface.vertexNormals[i].z = static_cast<float>(n[2]);
-    }
-  }
-  // Copy vertex colors
-  vtkDataArray *colors = mesh->GetPointData()->GetArray("Colors");
-  if (colors) {
-    double rgb[3];
-    surface.vertexColors.resize(mesh->GetNumberOfPoints());
-    for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-      colors->GetTuple(i, rgb);
-      surface.vertexColors[i].x = static_cast<float>(rgb[0]);
-      surface.vertexColors[i].y = static_cast<float>(rgb[1]);
-      surface.vertexColors[i].z = static_cast<float>(rgb[2]);
-    }
-  }
-  // Copy vertex UV coordinates
-  vtkDataArray *coords = mesh->GetPointData()->GetTCoords();
-  if (!coords)  coords = mesh->GetPointData()->GetArray("UV");
-  if (coords && (coords->GetNumberOfComponents() == 2 || coords->GetNumberOfComponents() == 3)) {
-    surface.vertexUV.resize(mesh->GetNumberOfPoints());
-    for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-      surface.vertexUV[i].u = static_cast<float>(coords->GetComponent(i, 0));
-      surface.vertexUV[i].v = static_cast<float>(coords->GetComponent(i, 1));
-    }
-  }
-  // Copy vertex labels
-  vtkDataArray *labels = mesh->GetPointData()->GetArray("Labels");
-  if (labels && labels->GetNumberOfComponents() == 1) {
-    surface.vertexLabels.resize(mesh->GetNumberOfPoints());
-    for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-      surface.vertexLabels[i] = static_cast<unsigned short>(labels->GetComponent(i, 0));
-    }
-  }
-  // Copy vertex attributes
-  vtkDataArray *scalars = mesh->GetPointData()->GetScalars();
-  if (!scalars) scalars = mesh->GetPointData()->GetArray("Attributes");
-  if (scalars && scalars->GetNumberOfComponents() == 1) {
-    surface.vertexAttributes.resize(mesh->GetNumberOfPoints());
-    for (vtkIdType i = 0; i < mesh->GetNumberOfPoints(); ++i) {
-      surface.vertexAttributes[i] = static_cast<float>(scalars->GetComponent(i, 0));
-    }
-  }
-  // Write .dfs file
-  return static_cast<int>(surface.writeDFS(fname));
-}
-
-// -----------------------------------------------------------------------------
-/// Write dataset points to TetGen .node file
-int WriteTetGenNode(ostream &os, vtkPolyData *polydata)
-{
-  vtkPointData *pd = polydata->GetPointData();
-  vtkDataArray *ar;
-  int nattributes = 0;
-  for (int i = 0; i < pd->GetNumberOfArrays(); ++i) {
-    nattributes += pd->GetArray(i)->GetNumberOfComponents();
-  }
-  os << polydata->GetNumberOfPoints() << " 3 " << nattributes << " 0\n";
-  double p[3];
-  streamsize precision = os.precision();
-  for (vtkIdType ptId = 0; ptId < polydata->GetNumberOfPoints(); ++ptId) {
-    os << (ptId + 1) << " ";
-    os.precision(8); // default TetGen tolerance is 1e-8
-    polydata->GetPoint(ptId, p);
-    os << " " << p[0] << " " << p[1] << " " << p[2];
-    os.precision(5);
-    for (int i = 0; i < pd->GetNumberOfArrays(); ++i) {
-      ar = pd->GetArray(i);
-      for (int j = 0; j < ar->GetNumberOfComponents(); ++j) {
-        os << " " << ar->GetComponent(ptId, j);
-      }
-    }
-    os << "\n";
-  }
-  os.precision(precision);
-  return os.fail() ? 0 : 1;
-}
-
-// -----------------------------------------------------------------------------
-/// Write dataset points to TetGen .node file
-int WriteTetGenNode(const char *fname, vtkPolyData *polydata)
-{
-  ofstream os(fname);
-  if (!os.is_open()) return false;
-  return WriteTetGenNode(os, polydata);
-}
-
-// -----------------------------------------------------------------------------
-/// Write polygonal dataset to TetGen .poly file
-int WriteTetGenPoly(const char *fname, vtkPolyData *polydata, const PointSet *holes)
-{
-  ofstream os(fname);
-  if (!os.is_open()) return 0;
-  os << "# part 1: nodes\n";
-  WriteTetGenNode(os, polydata);
-  vtkIdType npts, *pts;
-  vtkCellArray *verts  = polydata->GetVerts();
-  vtkCellArray *lines  = polydata->GetLines();
-  vtkCellArray *polys  = polydata->GetPolys();
-  vtkCellArray *strips = polydata->GetStrips();
-  int nfacets = 0;
-  if (verts ->GetNumberOfCells() > 0) nfacets += 1;
-  if (lines ->GetNumberOfCells() > 0) nfacets += 1;
-  if (polys ->GetNumberOfCells() > 0) nfacets += 1;
-  if (strips->GetNumberOfCells() > 0) nfacets += 1;
-  os << "\n# part 2: facets\n";
-  os << nfacets << " 0\n";
-  if (verts->GetNumberOfCells() > 0) {
-    os << "# verts\n";
-    os << verts->GetNumberOfCells() << "\n";
-    verts->InitTraversal();
-    while (verts->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (lines->GetNumberOfCells() > 0) {
-    os << "# lines\n";
-    os << lines->GetNumberOfCells() << "\n";
-    lines->InitTraversal();
-    while (lines->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (polys->GetNumberOfCells() > 0) {
-    os << "# polys\n";
-    os << polys->GetNumberOfCells() << "\n";
-    polys->InitTraversal();
-    while (polys->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (strips->GetNumberOfCells() > 0) {
-    os << "# strips\n";
-    os << strips->GetNumberOfCells() << "\n";
-    strips->InitTraversal();
-    while (strips->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  os << "\n# part 3: hole list\n";
-  if (holes) {
-    os << holes->Size() << "\n";
-    for (int i = 0; i < holes->Size(); ++i) {
-      const Point &p = holes->GetPoint(i);
-      os << (i+1) << " " << p._x << " " << p._y << " " << p._z << "\n";
-    }
-  } else {
-    os << "0\n";
-  }
-  os << "\n# part 4: region list\n";
-  os << "0\n";
-  return os.fail() ? 0 : 1;
-}
-
-// -----------------------------------------------------------------------------
-/// Write polygonal dataset to TetGen .smesh file
-int WriteTetGenSMesh(const char *fname, vtkPolyData *polydata, const PointSet *holes)
-{
-  ofstream os(fname);
-  if (!os.is_open()) return 0;
-  os << "# part 1: nodes\n";
-  WriteTetGenNode(os, polydata);
-  vtkIdType npts, *pts;
-  vtkCellArray *verts  = polydata->GetVerts();
-  vtkCellArray *lines  = polydata->GetLines();
-  vtkCellArray *polys  = polydata->GetPolys();
-  vtkCellArray *strips = polydata->GetStrips();
-  int nfacets = 0;
-  if (verts ->GetNumberOfCells() > 0) nfacets += verts ->GetNumberOfCells();
-  if (lines ->GetNumberOfCells() > 0) nfacets += lines ->GetNumberOfCells();
-  if (polys ->GetNumberOfCells() > 0) nfacets += polys ->GetNumberOfCells();
-  if (strips->GetNumberOfCells() > 0) nfacets += strips->GetNumberOfCells();
-  os << "\n# part 2: facets\n";
-  os << nfacets << " 0\n";
-  if (verts->GetNumberOfCells() > 0) {
-    verts->InitTraversal();
-    while (verts->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (lines->GetNumberOfCells() > 0) {
-    lines->InitTraversal();
-    while (lines->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (polys->GetNumberOfCells() > 0) {
-    polys->InitTraversal();
-    while (polys->GetNextCell(npts, pts)) {
-      os << npts << " ";
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  if (strips->GetNumberOfCells() > 0) {
-    strips->InitTraversal();
-    while (strips->GetNextCell(npts, pts)) {
-      os << npts;
-      for (vtkIdType i = 0; i < npts; ++i) os << " " << (pts[i] + 1);
-      os << "\n";
-    }
-  }
-  os << "\n# part 3: hole list\n";
-  if (holes) {
-    os << holes->Size() << "\n";
-    for (int i = 0; i < holes->Size(); ++i) {
-      const Point &p = holes->GetPoint(i);
-      os << (i+1) << " " << p._x << " " << p._y << " " << p._z << "\n";
-    }
-  } else {
-    os << "0\n";
-  }
-  os << "\n# part 4: region list\n";
-  os << "0\n";
-  return os.fail() ? 0 : 1;
-}
-
-// -----------------------------------------------------------------------------
-bool WriteOFF(const char *fname, vtkPolyData *polydata)
-{
-  ofstream ofs(fname);
-  ofs.precision(12);
-
-  ofs << "OFF\n";
-  ofs << polydata->GetNumberOfPoints() << " ";
-  ofs << polydata->GetNumberOfCells()  << " 0\n";
-
-  double p[3];
-  for (vtkIdType ptId = 0; ptId < polydata->GetNumberOfPoints(); ++ptId) {
-    polydata->GetPoint(ptId, p);
-    ofs << p[0] << " " << p[1] << " " << p[2] << "\n";
-  }
-
-  vtkIdType numPts, *ptIds;
-  polydata->BuildCells();
-  for (vtkIdType cellId = 0; cellId < polydata->GetNumberOfCells(); ++cellId) {
-    polydata->GetCellPoints(cellId, numPts, ptIds);
-    ofs << numPts;
-    for (vtkIdType i = 0; i < numPts; ++i) {
-      ofs << " " << ptIds[i];
-    }
-    ofs << "\n";
-  }
-
-  return !ofs.fail();
-}
-
-// -----------------------------------------------------------------------------
-bool WritePointSet(const char *fname, vtkPointSet *pointset, bool compress, bool ascii)
-{
-  vtkPolyData *polydata = vtkPolyData::SafeDownCast(pointset);
-  if (polydata) return WritePolyData(fname, polydata, compress, ascii);
-  const string ext = Extension(fname);
-  int success = 0;
-  if (ext.length() == 4 && ext.substr(0, 3) == ".vt" && ext != ".vtk") {
-    vtkSmartPointer<vtkXMLDataSetWriter> writer;
-    writer = vtkSmartPointer<vtkXMLDataSetWriter>::New();
-    SetVTKInput(writer, pointset);
-    writer->SetFileName(fname);
-    if (compress) writer->SetCompressorTypeToZLib();
-    else          writer->SetCompressorTypeToNone();
-    success = writer->Write();
-  } else {
-    vtkSmartPointer<vtkDataSetWriter> writer;
-    writer = vtkSmartPointer<vtkDataSetWriter>::New();
-    SetVTKInput(writer, pointset);
-    writer->SetFileName(fname);
-    if (ascii) writer->SetFileTypeToASCII();
-    else       writer->SetFileTypeToBinary();
-    success = writer->Write();
-  }
-  return (success == 1);
-}
-
-// -----------------------------------------------------------------------------
-bool WritePolyData(const char *fname, vtkPolyData *polydata, bool compress, bool ascii)
-{
-  const string ext = Extension(fname);
-  int success = 0;
-  if (ext == ".vtp") {
-    vtkSmartPointer<vtkXMLPolyDataWriter> writer;
-    writer = vtkSmartPointer<vtkXMLPolyDataWriter>::New();
-    SetVTKInput(writer, polydata);
-    writer->SetFileName(fname);
-    if (compress) writer->SetCompressorTypeToZLib();
-    else          writer->SetCompressorTypeToNone();
-    success = writer->Write();
-  } else if (ext == ".stl") {
-    vtkSmartPointer<vtkSTLWriter> writer;
-    writer = vtkSmartPointer<vtkSTLWriter>::New();
-    SetVTKInput(writer, polydata);
-    if (ascii) writer->SetFileTypeToASCII();
-    else       writer->SetFileTypeToBinary();
-    writer->SetFileName(fname);
-    success = writer->Write();
-  } else if (ext == ".ply") {
-    vtkSmartPointer<vtkPLYWriter> writer;
-    writer = vtkSmartPointer<vtkPLYWriter>::New();
-    SetVTKInput(writer, polydata);
-    if (ascii) writer->SetFileTypeToASCII();
-    else       writer->SetFileTypeToBinary();
-    writer->SetFileName(fname);
-    success = writer->Write();
-  } else if (ext == ".node") {
-    success = WriteTetGenNode(fname, polydata);
-  } else if (ext == ".poly") {
-    success = WriteTetGenPoly(fname, polydata);
-  } else if (ext == ".smesh") {
-    success = WriteTetGenSMesh(fname, polydata);
-  } else if (ext == ".dfs") {
-    success = WriteDFS(fname, polydata);
-  } else if (ext == ".off") {
-    success = WriteOFF(fname, polydata);
-  } else {
-    vtkSmartPointer<vtkPolyDataWriter> writer;
-    writer = vtkSmartPointer<vtkPolyDataWriter>::New();
-    SetVTKInput(writer, polydata);
-    writer->SetFileName(fname);
-    if (ascii) writer->SetFileTypeToASCII();
-    else       writer->SetFileTypeToBinary();
-    success = writer->Write();
-  }
-  return (success == 1);
-}
 
 // =============================================================================
 // VTK / MIRTK type conversion
